requirements.txt
requests~=2.31.0
openai~=0.27.8
python-dotenv~=1.0.0
flask~=2.2.2
redis~=4.6.0
tenacity~=8.2.2